+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.1] - 2020-04-04
+
+### Fixed
+
+- "Create Variable" now works well from variable declarations & JSX syntax
+
 ## [1.2.0] - 2020-04-04 - I handle async functions 🔮
 
 ### Added
@@ -39,7 +45,8 @@ async function doSomething() {}
 
 <!-- Links -->
 
-[unreleased]: https://github.com/nicoespeon/hocus-pocus/compare/1.2.0...HEAD
+[unreleased]: https://github.com/nicoespeon/hocus-pocus/compare/1.2.1...HEAD
+[1.2.1]: https://github.com/nicoespeon/hocus-pocus/compare/1.2.0...1.2.1
 [1.2.0]: https://github.com/nicoespeon/hocus-pocus/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/nicoespeon/hocus-pocus/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/nicoespeon/hocus-pocus/compare/5b3d351042d09ea26486598158069bce37b474b7...1.0.0